# Improve the throughput of SocketsHttpHandler's HTTP/1.1 connection pool