Closes #70098

The connection pool currently manages the list of available connections and the requests queue under a single lock.
As the number of cores and RPS rise, the speed at which the pool can manage connections becomes a bottleneck.

This PR brings the fast path (there are enough connections available to process all requests) down to a `ConcurrentStack.Push` + `ConcurrentStack.TryPop`.

---

<details>
<summary>Numbers for ConcurrentQueue</summary>

Numbers from https://github.com/dotnet/runtime/issues/70098#issuecomment-1969253633:

```
crank --config https://raw.githubusercontent.com/aspnet/Benchmarks/main/scenarios/httpclient.benchmarks.yml --scenario httpclient-kestrel-get --profile aspnet-citrine-lin --variable useHttpMessageInvoker=true --variable concurrencyPerHttpClient=256 --variable numberOfHttpClients=1 --server.framework net9.0 --client.framework net9.0 --json 1x256.json
crank --config https://raw.githubusercontent.com/aspnet/Benchmarks/main/scenarios/httpclient.benchmarks.yml --scenario httpclient-kestrel-get --profile aspnet-citrine-lin --variable useHttpMessageInvoker=true --variable concurrencyPerHttpClient=32 --variable numberOfHttpClients=8 --server.framework net9.0 --client.framework net9.0 --json 8x32.json
crank compare .\1x256.json .\8x32.json
```

| client      | 1x256        | 8x32        |         |
| ------------| ------------ | ----------- | ------- |
| RPS         |      693,873 |     875,814 | +26.22% |
| Patched RPS |      873,571 |     876,394 |  +0.32% |

This shows that before this PR, manually splitting load between multiple HttpClient instances can have a significant impact.
After the change, there's no more benefit to doing that as a single pool can efficiently handle the higher load.

YARP's http-http 100 byte scenario:

| load                    | yarp-base | yarp-patched |         |
| ----------------------- | --------- | ------------ | ------- |
| Latency 50th (ms)       |      0.73 |         0.68 |  -6.97% |
| Latency 75th (ms)       |      0.82 |         0.74 |  -9.82% |
| Latency 90th (ms)       |      1.03 |         0.89 | -13.39% |
| Latency 95th (ms)       |      1.41 |         1.18 | -16.41% |
| Latency 99th (ms)       |      2.87 |         2.68 |  -6.63% |
| Mean latency (ms)       |      0.83 |         0.76 |  -8.74% |
| Requests/sec            |   306,699 |      335,921 |  +9.53% |

In-memory loopback benchmark that stresses the connection pool contention: https://gist.github.com/MihaZupan/27f01d78c71da7b9024b321e743e3d88

Rough RPS numbers with 1-6 threads:

| RPS (1000s) | 1    | 2    | 3    | 4    | 5    | 6    |
| ------------| ---- | ---- | ---- | ---- | ---- | ---- |
| main        | 2060 | 1900 | 1760 | 1670 | 1570 | 1500 |
| patched     | 2150 | 2600 | 3400 | 3700 | 4100 | 4260 |


~~**Breaking change consideration**~~ - This is no longer relevant after switching to `ConcurrentStack`

~~While I was careful to keep the observable behavior of the pool as close as possible to what we have today, there is one important change I made intentionally:~~
- ~~The order in which we dequeue idle connections is changed from LIFO to FIFO (from a stack to a queue). This is because the backing store for available connections is now a `ConcurrentQueue`.~~
- ~~Where this distinction may be important is if a load drops for a longer period such that we no longer need as many connections. We would previously keep the overhead connections completely idle and eventually remove them via the idle timeout. With this change, we would keep cycling through all connections, potentially keeping more of them alive.~~
- ~~A slight benefit of that behavior may be that it makes it less likely to run into the idle close race condition (server closing an idle connection after we've started using it again).~~

</details>

See https://github.com/dotnet/runtime/pull/99364#issuecomment-2011014946 for `ConcurrentStack` results (current PR).