@MihaZupan 2024-03-21
Using a stack here is close enough (in benchmarks the collection is going to be close to empty all the time, so contention between the stack and queue is similar). I'll switch the PR to use that to avoid the [behavioral change](https://github.com/dotnet/runtime/pull/99364#:~:text=Breaking%20change%20consideration).
We can revisit it in the future with more idle eviction heuristics to get the last few % with a queue if needed.

It does mean an extra 32 byte allocation for each enqueue op sadly (+1 for #31911).

| load                    | yarp-main  | yarp-stack |         | yarp-queue |         |
| ----------------------- | ---------- | ---------- | ------- | ---------- | ------- |
| Latency 50th (ms)       |       0.73 |       0.69 |  -4.95% |       0.68 |  -5.91% |
| Latency 75th (ms)       |       0.81 |       0.75 |  -7.49% |       0.74 |  -8.97% |
| Latency 90th (ms)       |       0.99 |       0.88 | -10.98% |       0.85 | -14.40% |
| Latency 95th (ms)       |       1.31 |       1.13 | -13.99% |       1.05 | -20.00% |
| Latency 99th (ms)       |       2.83 |       2.60 |  -7.95% |       2.45 | -13.29% |
| Mean latency (ms)       |       0.82 |       0.76 |  -6.78% |       0.75 |  -8.59% |
| Requests/sec            |    312,857 |    335,444 |  +7.22% |    342,141 |  +9.36% |

| client             | client-main | client-stack |         | client-queue |         |
| ------------------ | ----------- | ------------ | ------- | ------------ | ------- |
| Requests           |  80,028,791 |  107,128,778 | +33.86% |  107,868,124 | +34.79% |
| Mean RPS           |     666,886 |      892,749 | +33.87% |      898,902 | +34.79% |

| Method    | Toolchain | Mean     | Error   | Ratio | Allocated | Alloc Ratio |
|---------- |-----------|---------:|--------:|------:|----------:|------------:|
| SendAsync |     main  | 517.0 ns | 4.27 ns |  1.00 |     552 B |        1.00 |
| SendAsync |     stack | 482.0 ns | 2.87 ns |  0.93 |     584 B |        1.06 |
| SendAsync |     queue | 471.1 ns | 1.37 ns |  0.91 |     552 B |        1.00 |